pyproject.toml
[project]
name = "auditeo-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Isuru Weerasingha", email = "info.isweerasingha@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "crewai>=1.11.0",
    "crewai-tools>=1.11.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pydantic-settings>=2.10.1",
    "python-json-logger>=4.0.0",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.6",
]


[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",
    "W",
    "F",
    "I",
    "C90",
    "B",
    "UP",
]
ignore = []

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
known-first-party = ["auditeo_ai"]
